Release v0.15.0
This is a minor release containing a few quality of life features. We
were originally planning to ship these with 1.0, but they're
specifically needed by some production users.

The biggest feature here is the ability to manually specify the `ON`
clause of a join, which is separate from the associations API. You can
also use this to join to tables which otherwise have no association
between them (but this will also require you to invoke
`enable_multi_table_joins!`).

In addition to the Diesel core team, 6 people contributed to this
release. A huge thank you to:

- Alex Kitchens
- Maciej Dziardziel
- Pyry Kontio
- Sunrin SHIMURA (keen)
- king6cong
- wangcong