Hide the token patterns we look for in `table!`
The generated documentation for the `table!` macro is really bad. Macros
document all of the patterns that they match on. For complex macros such
as this one, this is not useful information to the user.

This was maybe useful back in 0.1 when the docs looked like
https://docs.rs/diesel/0.1.0/diesel/macro.table!.html, but now it's just
a bunch of opaque macro soup that the user doesn't care about. The
actual documentation is buried more than two full page scrolls down.

This makes the documented token pattern as small as possible, so that
the actual documentation can be front and center.