Implement `FromSqlRow<Nullable<ST>, DB> for Option<T>` with blanket impl
This is one of those things that was definitely impossible, but I could
never quite remember why, and so every 2 months or so I look at it and
go "this should be possible as a blanket impl", implement it, and get a
compiler error that makes me remember why it's impossible. I have tried
this at least 4-5 times as part of that flow in the past.

I have no clue why it's possible this time. But it is. So yay?

This is something that is actually important to have as a blanket impl,
because a downstream crate *can not* write this impl themselves for
their types, as `Option` is not `#[fundamental]`